requirements.txt
python-telegram-bot==20.3
google-api-python-client==2.92.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
